FiniteVolumeMethod1D

Documentation for FiniteVolumeMethod1D.

FiniteVolumeMethod1D.FVMGeometry

FiniteVolumeMethod1D.FVMProblem

FiniteVolumeMethod1D.FVMGeometry — Type

FVMGeometry{T}

Definition of the geometry for a finite volume method problem.

Fields

mesh_points::T: The mesh points. Must be sorted.

spacings::T: The spacings between the mesh points.

volumes::T: The volumes of the cells defined by the mesh points.

Constructors

To construct the geometry, you can directly call the default constructor,

FVMGeometry(mesh_points, spacings, volumes)

or you can call the convenience constructor,

FVMGeometry(mesh_points)

which will compute the spacings and volumes for you.

See also FVMProblem.

FiniteVolumeMethod1D.FVMProblem — Type

FVMProblem{T,DF,DP,Dθ,RF,RP,Rθ,IC,FT}

Definition of an FVMProblem.

Fields

geometry::FVMGeometry{T}: The geometry of the problem.

diffusion_function::DF: The diffusion function.

diffusion_parameters::DP: The parameters for the diffusion function.

diffusion_theta::Dθ: The parameters for the diffusion function that are to be estimated.

reaction_function::RF: The reaction function.

reaction_parameters::RP: The parameters for the reaction function.

reaction_theta::Rθ: The parameters for the reaction function that are to be estimated.

initial_condition::IC: The initial condition.

initial_time::FT: The initial time.

final_time::FT: The final time.

Constructors

You can use the default constructor, but we also provide the constructor

FVMProblem(;
    geometry, 
    diffusion_function,
    diffusion_parameters = nothing,
    diffusion_theta = nothing,
    reaction_function = Returns(0.0),
    reaction_parameters = nothing,
    reaction_theta = nothing,
    initial_condition,
    initial_time = 0.0,
    final_time)

which provides some default values. Moreover, instead of providing geometry, you can use

FVMProblem(mesh_points; kwargs...)

which will construct geometry = FVMGeometry(mesh_points). The kwargs... are as above, except without geometry of course.

To solve the FVMProblem, just use solve as you would in DifferentialEquations.jl. For example,

sol = solve(prob, Tsit5(), saveat=0.1)